feat: 탈퇴사유 포함한 로직구현
