+++ src/components/molecules/RemoveContent/index.tsx
@@ -1,10 +1,15 @@
-import React, { ReactElement, useState, useEffect } from 'react';
+import React, { ReactElement, useState } from 'react';
+import { useHistory } from 'react-router-dom';
+import BottomBtn from '../../atoms/BottomBtn';
 import InputList from '../../atoms/InputList';
+import usePopAlarm from '../../../hooks/usePopAlarm';
 import * as S from './style';
 
 const RemoveContent = (): ReactElement => {
+  const history = useHistory();
   const [reasonListOpen, setReasonListOpen] = useState(false);
   const [reason, setReason] = useState('');
+  const [popAlarm] = usePopAlarm();
 
   const removeAccountList = [
     { id: 1, content: '선택해주세요.' },
@@ -22,25 +27,42 @@ const RemoveContent = (): ReactElement => {
     }
   };
 
+  const onNextClick = () => {
+    if (reason) {
+      // history.push('/mypage/removeAccount/2');
+      history.push({
+        pathname: '/mypage/removeAccount',
+        state: { reason },
+      });
+    } else {
+      popAlarm('이유를 선택해주세요');
+    }
+  };
+
   return (
-    <S.Container>
-      <S.H2>
-        빵덕을
-        <br />
-        <S.Span>탈퇴</S.Span> 하시나요?
-      </S.H2>
-      <S.P>탈퇴하시는 이유를 알려주세요</S.P>
-      <S.InputPositionPox>
-        <InputList
-          title="선택해주세요."
-          list={removeAccountList}
-          listOpen={reasonListOpen}
-          setListOpen={setReasonListOpen}
-          onClickItem={onReasonClick}
-          reason={reason}
-        />
-      </S.InputPositionPox>
-    </S.Container>
+    <>
+      <S.Container>
+        <S.H2>
+          빵덕을
+          <br />
+          <S.Span>탈퇴</S.Span> 하시나요?
+        </S.H2>
+        <S.P>탈퇴하시는 이유를 알려주세요</S.P>
+        <S.InputPositionPox>
+          <InputList
+            title="선택해주세요."
+            list={removeAccountList}
+            listOpen={reasonListOpen}
+            setListOpen={setReasonListOpen}
+            onClickItem={onReasonClick}
+            reason={reason}
+          />
+        </S.InputPositionPox>
+      </S.Container>
+      <S.BottomBox>
+        <BottomBtn content="다음" onClick={onNextClick} />
+      </S.BottomBox>
+    </>
   );
 };
