--- src/components/templates/RemoveAccountTemplate/style.ts
@@ -1,6 +1,0 @@
-import Styled from 'styled-components';
-
-export const BottomBox = Styled.div`
-  height:calc(100vh - 250px);
-  min-height: 466px;
-`;
